+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Disabled pedantic option on library compilation.
 - Defining `argus_init()` as a `static inline` function.
+- Fixed unexecuted validators in subcommand.
 
 ### Removed
 - Group description option.